requirements.txt
tensorflow-cpu==2.16.1
streamlit==1.35.0
numpy==1.24.0
pillow>=9.0.0